Clean up some bits
