+++ packages/outline-react/src/useOutlinePlainText.js
@@ -8,21 +8,41 @@
  */
 
 import type {OutlineEditor} from 'outline';
+import type {EventHandlerState} from 'outline-react/useOutlineInputEvents';
 
 import {useEffect, useMemo} from 'react';
+import {createTextNode} from 'outline';
 import useOutlineInputEvents from 'outline-react/useOutlineInputEvents';
-import {ParagraphNode} from 'outline-extensions/ParagraphNode';
+import {
+  createParagraphNode,
+  ParagraphNode,
+} from 'outline-extensions/ParagraphNode';
+
+function initEditor(editor: OutlineEditor): void {
+  editor.update((view) => {
+    const root = view.getRoot();
+
+    if (root.getFirstChild() === null) {
+      const text = createTextNode();
+      root.append(createParagraphNode().append(text));
+      text.select();
+    }
+  });
+}
 
 export default function useOutlinePlainText(
   editor: OutlineEditor,
   isReadOnly: boolean = false,
 ): {} | {onBeforeInput: (SyntheticInputEvent<EventTarget>) => void} {
-  const pluginState = useMemo(() => ({
-    isReadOnly: false,
-    richText: false,
-    compositionSelection: null,
-    isHandlingPointer: false,
-  }), []);
+  const pluginState: EventHandlerState = useMemo(
+    () => ({
+      isReadOnly: false,
+      richText: false,
+      compositionSelection: null,
+      isHandlingPointer: false,
+    }),
+    [],
+  );
 
   useEffect(() => {
     pluginState.isReadOnly = isReadOnly;
@@ -30,6 +50,7 @@ export default function useOutlinePlainText(
 
   useEffect(() => {
     if (editor !== null) {
+      initEditor(editor);
       return editor.addNodeType('paragraph', ParagraphNode);
     }
   }, [editor]);
