+++ packages/outline-react/src/useOutlineRichText.js
@@ -8,29 +8,47 @@
  */
 
 import type {OutlineEditor} from 'outline';
+import type {EventHandlerState} from 'outline-react/useOutlineInputEvents';
 
 import {useEffect, useMemo} from 'react';
+import {createTextNode} from 'outline';
 import useOutlineInputEvents from 'outline-react/useOutlineInputEvents';
 import {HeadingNode} from 'outline-extensions/HeadingNode';
 import {ListNode} from 'outline-extensions/ListNode';
 import {QuoteNode} from 'outline-extensions/QuoteNode';
 import {ParagraphNode} from 'outline-extensions/ParagraphNode';
 import {ListItemNode} from 'outline-extensions/ListItemNode';
+import {createParagraphNode} from 'outline-extensions/ParagraphNode';
+
+function initEditor(editor: OutlineEditor): void {
+  editor.update((view) => {
+    const root = view.getRoot();
+
+    if (root.getFirstChild() === null) {
+      const text = createTextNode();
+      root.append(createParagraphNode().append(text));
+      text.select();
+    }
+  });
+}
 
 export default function useOutlineRichText(
   editor: OutlineEditor,
   isReadOnly: boolean = false,
 ): {} | {onBeforeInput: (SyntheticInputEvent<EventTarget>) => void} {
-  const pluginState = useMemo(() => ({
-    isReadOnly: false,
-    richText: true,
-    compositionSelection: null,
-    isHandlingPointer: false,
-  }), []);
+  const eventHandlerState: EventHandlerState = useMemo(
+    () => ({
+      isReadOnly: false,
+      richText: true,
+      compositionSelection: null,
+      isHandlingPointer: false,
+    }),
+    [],
+  );
 
   useEffect(() => {
-    pluginState.isReadOnly = isReadOnly;
-  }, [isReadOnly, pluginState]);
+    eventHandlerState.isReadOnly = isReadOnly;
+  }, [isReadOnly, eventHandlerState]);
 
   useEffect(() => {
     if (editor !== null) {
@@ -42,6 +60,7 @@ export default function useOutlineRichText(
         ParagraphNode,
       );
       const removeListItemType = editor.addNodeType('listitem', ListItemNode);
+      initEditor(editor);
 
       return () => {
         removeHeadingType();
@@ -53,5 +72,5 @@ export default function useOutlineRichText(
     }
   }, [editor]);
 
-  return useOutlineInputEvents(editor, pluginState);
+  return useOutlineInputEvents(editor, eventHandlerState);
 }
